Code correction after hand testing pile 2 : HomepageActivity, Workplace persistence (SharedPref and Firestore), Map (init on observe), RestaurantList (init on observe + workplace Btn), ColleaguesList (init on observe), RestaurantDetails (init on observe + likeBtn bugfix), FirebaseHandler (logic enhanced).

+++ app/src/main/java/fr/joudar/go4lunch/domain/dto/MapApiPlaceDetailsResponse.java
@@ -27,7 +27,7 @@ public class MapApiPlaceDetailsResponse {
             resultPlace = new Place(
                     result.placeId,
                     result.name,
-                    result.rating.doubleValue(),
+                    result.rating,
                     result.formattedAddress,
                     toPhotos(),
                     result.website,
